+++ packages/server/src/sdk/app/rows/search/internal/internal.ts
@@ -1,90 +1,30 @@
 import { context, HTTPError } from "@budibase/backend-core"
-import { PROTECTED_INTERNAL_COLUMNS } from "@budibase/shared-core"
-import env from "../../../../environment"
-import { fullSearch, paginatedSearch } from "./utils"
-import { getRowParams, InternalTables } from "../../../../db/utils"
+import env from "../../../../../environment"
+import { getRowParams, InternalTables } from "../../../../../db/utils"
 import {
   Database,
   DocumentType,
   Row,
-  RowSearchParams,
-  SearchResponse,
-  SortType,
   Table,
   TableSchema,
-  User,
 } from "@budibase/types"
-import { getGlobalUsersFromMetadata } from "../../../../utilities/global"
-import { outputProcessing } from "../../../../utilities/rowProcessor"
+import { outputProcessing } from "../../../../../utilities/rowProcessor"
 import {
   csv,
   Format,
   json,
   jsonWithSchema,
-} from "../../../../api/controllers/view/exporters"
-import * as inMemoryViews from "../../../../db/inMemoryView"
+} from "../../../../../api/controllers/view/exporters"
+import * as inMemoryViews from "../../../../../db/inMemoryView"
 import {
   getFromDesignDoc,
   getFromMemoryDoc,
   migrateToDesignView,
   migrateToInMemoryView,
-} from "../../../../api/controllers/view/utils"
-import sdk from "../../../../sdk"
-import { ExportRowsParams, ExportRowsResult } from "./types"
-import pick from "lodash/pick"
-import { breakRowIdField } from "../../../../integrations/utils"
-
-export async function search(
-  options: RowSearchParams,
-  table: Table
-): Promise<SearchResponse<Row>> {
-  const { tableId } = options
-
-  const { paginate, query } = options
-
-  const params: RowSearchParams = {
-    tableId: options.tableId,
-    sort: options.sort,
-    sortOrder: options.sortOrder,
-    sortType: options.sortType,
-    limit: options.limit,
-    bookmark: options.bookmark,
-    version: options.version,
-    disableEscaping: options.disableEscaping,
-    query: {},
-  }
-
-  if (params.sort && !params.sortType) {
-    const schema = table.schema
-    const sortField = schema[params.sort]
-    params.sortType =
-      sortField.type === "number" ? SortType.NUMBER : SortType.STRING
-  }
-
-  let response
-  if (paginate) {
-    response = await paginatedSearch(query, params)
-  } else {
-    response = await fullSearch(query, params)
-  }
-
-  // Enrich search results with relationships
-  if (response.rows && response.rows.length) {
-    // enrich with global users if from users table
-    if (tableId === InternalTables.USER_METADATA) {
-      response.rows = await getGlobalUsersFromMetadata(response.rows as User[])
-    }
-
-    if (options.fields) {
-      const fields = [...options.fields, ...PROTECTED_INTERNAL_COLUMNS]
-      response.rows = response.rows.map((r: any) => pick(r, fields))
-    }
-
-    response.rows = await outputProcessing(table, response.rows)
-  }
-
-  return response
-}
+} from "../../../../../api/controllers/view/utils"
+import sdk from "../../../../../sdk"
+import { ExportRowsParams, ExportRowsResult } from "../types"
+import { breakRowIdField } from "../../../../../integrations/utils"
 
 export async function exportRows(
   options: ExportRowsParams
@@ -123,15 +63,12 @@ export async function exportRows(
 
     result = await outputProcessing<Row[]>(table, response)
   } else if (query) {
-    let searchResponse = await search(
-      {
-        tableId,
-        query,
-        sort,
-        sortOrder,
-      },
-      table
-    )
+    let searchResponse = await sdk.rows.search({
+      tableId,
+      query,
+      sort,
+      sortOrder,
+    })
     result = searchResponse.rows
   }
